server vnc, dockerfile streamlining, vm script
Adds functionality for the server to create a vnc container, grabbing an extra two contiguous ports (because that's how VNC apparently works) and binding them to the internal VNC ports 5900 (base) and 5901 (screen one).

Streamlines Dockerfiles to reduce size (by a ton, actually).

Adds a script to create a GCP VM that gets all dependencies, builds the images from the two dockerfiles, and starts the server. Note that the server won't work properly until the builds finish. See the comment in the file for why we aren't using docker save/load. Dockerhub would probably be an acceptable substitute, but I'm lazy and have to mess with the Dockerfiles all the time.